// src/shared/api/generated/schemas/workspacesMemberResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik Pro API
 * HTTP API сервиса smethick-pro.
 * OpenAPI spec version: 1.0
 */
import type { WorkspacesMemberRoleResponse } from './workspacesMemberRoleResponse';

export interface WorkspacesMemberResponse {
  created_at?: string;
  email?: string;
  id?: string;
  name?: string;
  phone?: string;
  position?: string;
  role?: WorkspacesMemberRoleResponse;
  surname?: string;
  telegram?: string;
  updated_at?: string;
  user_id?: string;
  workspace_id?: string;
}